macOS: fixes required to build and run DKR-R 1.0.4 on Apple Silicon
Corresponding source for the DKR-R-1.0.4-macOS-arm64 release. The runtime
links GPL-3.0 N64ModernRuntime, so the binary's complete source must be
available; upstream 1.0.4 alone does not build on macOS.

Build tooling:
- apply-dependency-patches.sh tested .git with is_dir(), which is false for a
  git submodule (a .git file). N64Recomp is the only submodule dependency, so
  its two patches were silently skipped and recompilation failed with
  "Could not find entrypoint function".
- Prepare-DKR-Runtime-macOS.sh expanded an empty array under `set -u`, which
  bash 3.2 (the version macOS ships) treats as unbound.

Missing macOS platform support:
- runtime_platform.cpp had no __APPLE__ branch for the window handle. RT64
  casts RenderWindow::view to CA::MetalLayer*, so it must receive the layer,
  not the NSView. The layer is rebuilt at launcher handoff and leftover
  sublayers removed, otherwise the launcher's last frame re-composites over
  the running game on focus loss.
- runtime_support.cpp used <sys/sysinfo.h> and /proc; now sysctl.
- rom_revision.cpp and runtime_texture_packs.cpp streamed
  file_time_type::rep, which is __int128 on libc++.
- runtime_hud_layout.cpp used std::atomic<std::shared_ptr<T>>, which libc++
  does not implement.
- runtime_ui.cpp did not set the SDL_Renderer scale, so the launcher drew at
  1/backingScaleFactor on Retina and the pointer missed the UI.

Not macOS-specific:
- gekko_direct_adapter.cpp locked a mutex destroyed earlier at exit() by a
  global in another translation unit. Undefined behavior everywhere; libc++
  diagnoses it and aborts.

Also adds scripts/Bundle-macOS-Redistributable.sh, which vendors SDL2/SDL3
into the bundle so it runs without Homebrew, and macos-patches/ holding the
same changes as a standalone series (including the RT64 Retina fix, which
applies to extern/ after dependency preparation).

+++ runtime-recomp/src/game/runtime_texture_packs.cpp
@@ -104,7 +104,7 @@ std::string FileSignature(const std::filesystem::path& path) {
     const auto modified = std::filesystem::last_write_time(path, error);
     if (error) return "missing";
     return std::to_string(size) + ":" +
-        std::to_string(modified.time_since_epoch().count());
+        std::to_string(static_cast<long long>(modified.time_since_epoch().count()));
 }
 
 std::string PackFingerprint(const std::filesystem::path& path,
@@ -113,7 +113,8 @@ std::string PackFingerprint(const std::filesystem::path& path,
     std::error_code error;
     const auto modified = std::filesystem::last_write_time(path, error);
     const std::string directory_time = error
-        ? "missing" : std::to_string(modified.time_since_epoch().count());
+        ? "missing"
+        : std::to_string(static_cast<long long>(modified.time_since_epoch().count()));
     return "directory:" + directory_time + ":database:" +
         FileSignature(path / "rt64.json") + ":report:" +
         FileSignature(path / "dkr-r-rice-import.json");
